+++ olca-ilcd/src/test/java/org/openlca/ilcd/io/UnitGroupBagTest.java
@@ -61,7 +61,7 @@ public class UnitGroupBagTest {
 	@Test
 	public void testGetClasses() {
 		Class clazz = bag.getSortedClasses().get(0);
-		assertEquals(0, clazz.level.intValue());
+		assertEquals(0, clazz.level);
 		assertEquals("Technical unit groups", clazz.value.trim());
 	}
